# ec_backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/admin/login": {
      "post": {
        "operationId": "AuthController_adminLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "Administrator successfully logged in."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Please check the request. Either the user or the password are incorrect."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "User successfully logged in!"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Please check the request. Either the user or the password are incorrect."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "We have sent you an email with a link to confirm your email address."
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "example": "The user already exists."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/registered": {
      "patch": {
        "operationId": "AuthController_emailConfirmed",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "title": "Token",
              "example": "#hd)@jl4k019d0KKL",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "User successfully logged in!"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The token is expired."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "get": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "User successfully logged out."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset/password": {
      "patch": {
        "operationId": "AuthController_updatePassword",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updatePasswordDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "Your password has been successfully updated. If you were logged in, for security reasons, you have been logged out."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The token is invalid."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "post": {
        "operationId": "AuthController_generatePasswordResetToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateResetTokenDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "If the email is correctly registered, you should have received a link to reset your password."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset/email": {
      "patch": {
        "operationId": "AuthController_updateEmail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "Your email has been successfully updated. For security reasons, you have been logged out."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Invalid email. Check the request."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "post": {
        "operationId": "AuthController_generateEmailResetToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateResetTokenDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "If the email is correct, you should have received a link to reset your email."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "English Conversations API",
    "description": "REST API for managing English Conversations' data and processes.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "example": "john.doe@mail.com"
          },
          "password": {
            "type": "string",
            "title": "IsTest",
            "example": "#i8NNhl10@%"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "updatePasswordDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "example": "john.doe@mail.com"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "example": "Em85*&kl31rD"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "GenerateResetTokenDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "example": "john.doe@mail.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  }
}